Bug 1743995 - DNS cache can not get TTL from DNS-over-HTTPS r=necko-reviewers,kershaw
If mTTL is initialized to UINT32_MAX then the update condition when
adding a new IP via DOHResp::Add will never be triggered (a TTL can
never be larger than UINT32_MAX).
So it is better to have it be initialized to 0.

Differential Revision: https://phabricator.services.mozilla.com/D132697

UltraBlame original commit: 839eb1957f913ddfbebbeb1e56d9846140fb8512