Bug 1744286 - Use a bitmap to store free committed arenas in the chunk header r=sfink
This replaces use of the free arena linked list with a bitmap. The next free
arena is found using bit operations. The number of arenas in a chunk is 256,
which is 32 bytes and should fit in a cache line.

This patch greatly simplifies the releveant code. It also give a 5% improvement
on the splay benchmark on Linux. I'm not sure exactly why this is but perf
indicates an increased number of context switches so it's probably locking
behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D132954

UltraBlame original commit: 8c8034a8ab0f7aee7534f84ff2fbab5761b010a8