Bug 1741708 - [devtools] Fix exception message filtering for Web extension debugging. r=jdescottes
The original report isn't against Nightly and so is before bug 1675456 landed.
It means that on current release, we are still going throught the WebConsole actor
and use ConsoleServiceListener, which isn't filtering anything:
this.window is null because targetActor.isRootActor=true:
 https://searchfox.org/mozilla-central/rev/65d4d3399afa79c8de5a0cc11752d2ba7c31edc1/devtools/server/actors/webconsole/listeners/console-service.js#72-102)
=> on release we log all exception, like the MBT.

Whereas on Nightly, bug 1675456 kicks in and we are going using ErrorMessageWatcher.
This time, this class filter out message and only include the one related to the target's current top window.
So it is slightly better, but we may miss messages of other documents of the add-on.
For example, if a popup window is opened after having opened the DevTools.

This patch addresses the watcher class codepath. Only tests should be using the console actor and ConsoleServiceListener.
We may later drop this code once we start spawning one target for each individual document
used by an add-on. For now, the WebExtensionTargetActor is a unique target that debugs all extension's documents.

Differential Revision: https://phabricator.services.mozilla.com/D132230

UltraBlame original commit: ce5f81f38c5bc64909add59270224dbc0b9155ca