Bug 1680846 - Part 1 - requestStorageAccess should always result in a permission prompt for 3p requests and no prior 1p interaction r=timhuang
storageAccessAPI permission is now required for an autogrant. This increases the number of prompts users will see, but not from any pages
interacted with as a first-party. Combined with the resolution of Bug 1680844, this will also mean that a third-party that has an
embed-only use pattern will display a prompt only once, if it is accepted.

This is implemented by checking for the storageAccessAPI permission when determining whether or not to autogrant.

Differential Revision: https://phabricator.services.mozilla.com/D127276

UltraBlame original commit: d916d9ac25e9810003dd86c6224deeb5034f175a