Bug 1738971 - Part 2. Add WorkerDocumentListener to get visibility events. r=dom-worker-reviewers,gfx-reviewers,jgilbert,smaug
requestAnimationFrame callbacks are supposed to be delayed if the
worker's owning document is no longer visible. This patch adds the
ability for a WorkerGlobalScope to listen to changes in the visiblility
state of the Document which owns the worker.

Differential Revision: https://phabricator.services.mozilla.com/D130263

UltraBlame original commit: d9aef3e9797ef1052b6e600c7d9922ec8b85664a