Bug 1732358 - Part 1: Override required_prefs in test_fission_autostart.py, r=whimboo,mccr8
Unfortunately we currently always configure the "fission.autostart" pref when
running tests, meaning that the test_fission_autostart.py test is always being
skipped on all platforms. This change force-disables the required pref when
running the test, rather than skipping the test if it is set.

It would be nice to handle this in a less hacky way, but this was the first way
I could think of.

Differential Revision: https://phabricator.services.mozilla.com/D133005

UltraBlame original commit: f341c2c8bbee3fe84c70ad0a0b8524d106120663