Bug 1744468. Make drawSnapshot request high quality scaling by default. r=aosmond
Most of the time this is just going to going to request the same frame that we used when painting to the screen. So it should be higher quality and faster in the common case.

This isn't the default for drawWindow or drawSnapshot. I know that it would be a waste to request high quality scaling without sync decode because if that triggers new decoding by the time the frame is ready it's too late to use it. drawSnapshot only supports syncdecoding so far, so that's not an issue, and we could disable this is async decoding was requested in the future.

Anything else I'm missing?

Differential Revision: https://phabricator.services.mozilla.com/D133567

UltraBlame original commit: 20c0dc7a0122e5110076da1e9fb138934c30c737