Bug 1744468. Change annotation for gfx/tests/reftest/1724901-2.html. r=aosmond
This test has a downscaled image, that is inside of an OOP iframe, and the iframe has a large transform scale to counter act the downscaled image inside the iframe. drawSnapshot does not take into account into scaling that happens to OOP iframes. If we don't pass the high quality scaling flag then we return the image as the intrinsic scale and we have a much better rendering here. If we pass the high quality scale we return the downscaled image, which we then try to upscale, and we get a worse rendering then before the patches of this bug.

So this is an edge that we will be making worse. If it's important then we should fix drawSnapshot to handle this.

Differential Revision: https://phabricator.services.mozilla.com/D133569

UltraBlame original commit: 36dc359531fc1993d990b04f109bb41b258dac0b