Bug 1728327 - Track ExtensionEventListener a background service worker registered by the time it is fully loaded. r=baku
This patch includes a proposed approach to keep track of the WebExtensions API event listeners
subscribed synchronously while the background service worker script was being loaded and executed,
because this are the listeners that we can assume to be available right after we spawn a worker
and be able to handle the API event that triggered the worker to be spawned.

The information about the "listeners subscribed synchronously while the worker script is being
loaded and executed" is then used by the ExtensionBrowser::HasWakeupEventListener method,
which will be called as part of handling the nsIServiceWorkerManager.wakeForExtensionAPIEvent
call.

Differential Revision: https://phabricator.services.mozilla.com/D130758

UltraBlame original commit: 199796d1623becd174b2317bdb39871c58a8e20e