Bug 1746095 - Avoid crashing when we get memory pressure events with OffscreenCanvas. r=jgilbert
We will implement this properly in bug 1746260, but for now, let's spare
users with OffscreenCanvas turned on, who have low memory, a content
process crash. Handling it is somewhat complicated because we need to
shuttle over the message to the worker.

Differential Revision: https://phabricator.services.mozilla.com/D133939

UltraBlame original commit: 2cb4a2403bd0856cdf4da19467bde16908f49e8d