Bug 1742993 - Non-fatally deny scheduling parameter operations on other threads in the GMP sandbox. r=gcp
These syscalls operate on a given thread ID, or 0 to indicate the
calling thread; the seccomp-bpf policy can allow the fixed value 0, and
a SIGSYS trap handler can check for the current thread's ID and remap it
to 0, but we don't have a safe way to check if it's another thread of
the current process (which is the policy we'd like to have).  If they
weren't filtered at all, they could act on any thread of any process
owned by the same user, which is not ideal.

Currently the GMP process policy allows them for the calling thread
and treats them as unexpected syscalls otherwise (crashing by default
on Nightly, else ENOSYS), while the content process policy allows them
without filtering (bug 1413313) and we need also them in the RDD process
(this bug).  The direction we're going to take is to allow them on the
current thread and fail with EPERM otherwise.

This patch, therefore, changes the default behavior to EPERM and moves
the code into the policy superclass so that other process types can share it.

Differential Revision: https://phabricator.services.mozilla.com/D133711

UltraBlame original commit: 93992f9b75235cca8f169a6d46132e303cb31a71