Bug 1745949: Clone before truncating r=jandem,nbp
Before truncating an instruction, RangeAnalysis sometimes creates a cloned copy for use in the bailout path. The original design intent was that the cloned copy would not be truncated. However, we currently call `needTruncation` to decide whether to truncate a node, then clone it, then call `truncate`. Somewhat counterintuitively, most implementations of `needTruncation` modify the truncation kind for the node. Because this happens before we clone, the cloned copy ends up being truncated.

In this particular fuzzbug, edgecase analysis sees that the only consumer of a multiplication by -1 is a truncated addition, and removes the negative zero check. If we clone the MAdd before truncating it, then edgecase analysis does the right thing.

To accomplish this, for each MIR instruction that implements `needTruncation`/`truncate`, I made `needTruncation` a const function and moved any mutation into `truncate` itself. This ensures that the cloned node will be an unmodified copy of the original.

I took a quick look at Octane, and this doesn't appear to have affected our score.

Differential Revision: https://phabricator.services.mozilla.com/D133955

UltraBlame original commit: 4a3fbe0d0746848a6593a5b2e30ea0fc3cf56ca0