Bug 1739695 [wpt PR 31526] - webauthn: support the minPinLength extension., a=testonly
Automatic update from web-platform-tests
webauthn: support the minPinLength extension. (#31526)

This extension[1] allows a configured site to learn the minimum PIN
length set on an authenticator. This is for enterprise environments that
need to enforce PIN length policies.

The CTAP 2.1 specification allows a) the minimum PIN length to be
increased from the usual four bytes, b) a domain to be configured that
is authorised to request a report on the current minimum, and c) this
extension to plumb that data back to the site.

Thus, if a security key is configured to allow, say, example.com to
learn the minimum PIN length, and example.com creates a credential to
learn that the PIN length is $x, then it knows that the PIN length
enforced when that credential is exercised is never less than that. (The
only way to reduce the minimum is to reset the authenticator, but that
wipes all credentials.)

Chromium does not provide any mechanism for configuring the allowed
domain. That has to be done manually.

Intent to Ship thread: https://groups.google.com/a/chromium.org/g/blink-dev/c/VnXR-U3jROc

[1]
https://fidoalliance.org/specs/fido-v2.1-ps-20210615/fido-client-to-authenticator-protocol-v2.1-ps-20210615.html#sctn-minpinlength-extension

BUG=1259760

Change-Id: Ic4230b9986c8a4215e1f80da9616a571d7b2b7c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3256056
Reviewed-by: Adam Langley <aglchromium.org>
Reviewed-by: Mike West <mkwstchromium.org>
Reviewed-by: Ken Buchanan <kenrbchromium.org>
Reviewed-by: Martin Kreichgauer <martinkrgoogle.com>
Commit-Queue: Adam Langley <aglchromium.org>
Cr-Commit-Position: refs/heads/main{#946127}

Co-authored-by: Adam Langley <aglchromium.org>
--

wpt-commits: 6e3c94a63090cd0e38761550495c22f2fa79d5f0
wpt-pr: 31526

UltraBlame original commit: 2f6b7901372d420e248b016b97d96c7316d6d32c