Bug 1746852 - Add `scriptId` to `registeredScriptIds` to be able to unregister all user scripts when context is being closed. r=rpl
We register a function that is called when the context is getting closed
but this function relies on a set of script IDs that is always empty,
because we don't add any IDs to it when we register a user script. This
patch adds script IDs of registered user scripts to this set.

Differential Revision: https://phabricator.services.mozilla.com/D134278

UltraBlame original commit: 5eb78be7a83caa08805ced60b3cdca1adf3d5aed