Bug 1730843 - Part 15 - Implement support for R&T as Map/Set keys r=jandem
When preparing records, tuples and boxes to be used as a Map/Set key, we
recursively atomize the strings that they contain. This makes it
possible to compute a hash for the whole structure, and to compare them
with an infallible function.

ToDo: Figure out where to store the isAtomized flag for tuples

Differential Revision: https://phabricator.services.mozilla.com/D125660

UltraBlame original commit: 5820e13028f9b1a2930b00b6b6a260dc9271e41c