Bug 1746893 - Relax AudioDriftCorrection's thread limitation r=padenot,pehrsons
By the comment in `AudioDriftCorrection`: "The construction can happen
in any thread", the `AudioDriftCorrection` class is ought to be created
in any thread, but it's not true actually. The fact that ctor calls
`Preferences::GetInt` makes it can only be called on main-thread, which
is the thread `Preferences::GetInt` should be.

To make `AudioDriftCorrection` work as what it's expected, the
preference getter should be moved to the `AudioDriftCorrection`'s caller
rather than `AudioDriftCorrection'`s ctor.

Differential Revision: https://phabricator.services.mozilla.com/D122297

UltraBlame original commit: 0714fcb047accbe6a255811801b30634709dbcff