Bug 1728436 - Vendor the midir crate r=padenot
This is as close to usptream as currently possibly. Only a few changes were
done to the dependencies: the wasm target was removed and the coremidi
dependency was updated to pick up a more recent version so that we don't need
to vendor separate versions of the core-foundation and core-foundation-sys
crates.

This vendors the following crates:
* alsa-sys
* alsa
* coremidi
* coremidi-sys
* memalloc
* midir

Overall this adds ~30K lines of code, over half of which is in the alsa
bindings alone.

Differential Revision: https://phabricator.services.mozilla.com/D124640

UltraBlame original commit: 791abc86e79913d8c72767de0c072c5e9fed7354