Bug 1730712: Abstract site behaviour from requirements.py r=ahal
There are two sites that are allowed to define their dependencies
in a flexible way: the `mach` and `build` sites.

This is because these are the only two sites that _may_ have
to operate without `pip install`-ing any packages, and instead having
to be compatible with the packages installed to the system.

Due to this required compatibility, allowing flexibility to these sites
allows flexibility downstream.

Anyways, this patch isn't about that - that behaviour has already
landed. This patch is about tweaking `requirements.py` so that
*it* doesn't care about specific sites, but rather only cares about
if it should assert `only_strict_requirements` or not. Accordingly,
the helpful "not all packages are pinned" error message is moved
to `site.py`, where it belongs.

Differential Revision: https://phabricator.services.mozilla.com/D132082

UltraBlame original commit: 1d9613af189689f85c0e973d6b89fcf10e8c3771